Infectious_CDM-public/examples/requirements.txt
pandas==1.4.4
numpy==1.23.5
PyYAML==6.0